+++ rwaClient/src/AppDelegate.swift
@@ -319,8 +319,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
        }
 
     func applicationWillTerminate(_ application: UIApplication) {
-        
-        // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
+        // Synchronous stop: no fade, immediate teardown.
+        (currentSceneController as? SecondViewController)?.stopGameNow()
     }
 }